#98 Fix expand and buitins for norminette
